requirements.txt
torch
torchvision
pandas
tqdm
Pillow
pyinaturalist
pytorch-lightning
numpy
scipy